requirements.txt
ollama
gitpython
pyyaml
python-dotenv
requests